requirements.txt
alembic==1.13.3
amqp==5.2.0
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.6.0
appdirs==1.4.4
bidict==0.23.1
billiard==4.2.1
blinker==1.8.2
celery==5.3.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-log==0.4.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
deprecation==2.1.0
docutils==0.21.2
dotty-dict==1.3.1
Flask==2.3.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-RESTful==0.3.10
Flask-Script==2.0.6
Flask-SocketIO==5.3.4
Flask-SQLAlchemy==3.0.5
gitdb==4.0.11
GitPython==3.1.43
gotrue==1.3.1
greenlet==3.1.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.10
importlib_metadata==8.5.0
invoke==1.7.3
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.4
keyboard==0.13.5
keyring==25.4.1
kombu==5.4.2
Mako==1.3.5
MarkupSafe==3.0.1
more-itertools==10.5.0
nh3==0.2.18
packaging==24.1
pkginfo==1.11.1
postgrest==0.10.7
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
pyee==8.2.2
Pygments==2.18.0
PyJWT==2.6.0
pynput==1.7.7
pyppeteer==1.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-engineio==4.9.1
python-gitlab==3.15.0
python-semantic-release==7.33.2
python-socketio==5.11.4
pytz==2024.2
pywin32-ctypes==0.2.3
readme_renderer==44.0
realtime==1.0.0
redis==4.5.5
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==1.5.0
semver==2.13.0
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.19
storage3==0.5.3
StrEnum==0.4.15
supabase==1.0.3
supafunc==0.2.2
tomlkit==0.13.2
tqdm==4.66.5
twine==3.8.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
websockets==10.4
Werkzeug==2.3.6
wheel==0.44.0
wsproto==1.2.0
zipp==3.20.2
